Move common functions into proteus-common
